Move update_permissions_link to new helper file
I want to use this helper method in other application tables so I'm
first moving it to a more generically-named helper.
